rename `store_file.rs` to `store.rs`
